requirements.txt
timm==0.8.15dev0 
mmselfsup 
openpyxl 
imgaug 
numba 
tensorboard==2.12.0
fvcore 
accimage 
Ninja
numpy==1.24.1
albumentations==1.3.0
annotator==0.14.2
datasets==2.14.4
einops==0.3.0
gradio==3.40.1
omegaconf==2.1.1
open_clip_torch==2.0.2
opencv_contrib_python==4.3.0.36
opencv_python==4.5.1.48
opencv_python_headless==4.5.1.48
pandas==1.3.4
Pillow==9.4.0
pytorch_lightning==1.5.0
safetensors==0.2.7
scikit_learn==1.2.2
scipy==1.9.1
scikit-image==0.20.0
tabulate==0.9.0
torchmetrics==0.6.0
tqdm==4.65.0
transformers==4.19.2
xformers==0.0.18
taming-transformers
matplotlib==3.7.1